mock companion app connection
