+++ deployables/extension/package.json
@@ -17,7 +17,7 @@
     "zip": "mv public zodiac-pilot && zip -vr zodiac-pilot.zip zodiac-pilot/ -x \"*.DS_Store\" && mv zodiac-pilot public",
     "predev": "pnpm clean",
     "dev": "dotenvx run -- pnpm build",
-    "test": "vitest",
+    "test": "dotenvx run -- vitest",
     "test:e2e": "PW_CHROMIUM_ATTACH_TO_OTHER=1 playwright test --headed",
     "test:e2e:ui": "pnpm test:e2e --ui",
     "check-types": "tsc --noEmit",
