+++ deployables/extension/src/panel/companion/index.ts
@@ -3,3 +3,4 @@ export {
   useCompanionAppUrl,
 } from './CompanionAppSupport'
 export { getFeatures } from './getFeatures'
+export { getUser } from './getUser'
